Add files via upload
Hardware Components
Buttons:
1.Button: When pressed for 3 seconds, it is used to set the hour, and when pressed and pulled, it
switches between time units. It is also used to increase the values of the current unit in any setting
condition.
2.Button: When pressed once, it turns the alarm on and off, and when pressed for 3 seconds, it sets
the alarm time.
3.Button: Allows switching between temperature units.
4.Button: It is used to snooze the alarm.
Resistors:
Pull-Down Resistors: It is to prevent the Arduino pins from being short-circuited with gnd and causing
instability when the button is pressed.
Potentiometer: It is used to adjust the brightness of the LCD screen.
Sensors and Other Elements:
DHT11: It is a temperature and humidity sensor. Although there is a very small margin of error, it
gives 98% correct results. It communicates with the Arduino thanks to the signals coming from the
Out pin.
16x2 LCD Screen: The screen where we print the time, temperature and alarm data.
Software Components
In this project; An alarm clock with temperature display was designed using DHT and LiquidCrystal
libraries.
Operations inside the void setup() function:
1. Starting DHT11 communication
2. Defining some special characters for the LCD screen and adjusting the LCD screen
3. Setting pin functions
Operations inside the void loop() function:
1. Reading the potentiometer value and adjusting the LCD display light
2. Defining the functions required for clock and alarm settings
3. Defining the primary and secondary functions of the buttons
4. Calling the DisplayTemp() function to display the temperature value